fix: redirect to aliascontent changelist after deleting alias
When deleting an Alias via the toolbar when versioning is not enabled redirect
to the AliasContent admin changelist instead of custom LIST_ALIASES_URL_NAME
view as this will be removed.
Updates toolbar test so that it is always run, with versioning mocked to be
disabled.
Updates more existing tests to use the AliasContent list view instead of the
LIST_ALIASES_URL_NAME view as this will be removed.